requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
celery==5.4.0
redis==5.0.8
sse-starlette==2.1.3
python-multipart==0.0.9
minio==7.2.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0


